properties: factor out a toggle_playlist functions for menu and panel
